#127/imp/obsolete CreatePost replaced
